RF: `actor.py` file refractored into multiple files
- Files include init, curved, polyhedron, planar, core, slicer, surface, bio.
- Reduces the load on low RAM machines, with increasing extension for productivity.
- Slicer actor renamed to `data_slicer`
