+++ fury/actor/__init__.py
@@ -1,0 +1,5 @@
+"""Init file for visualization package."""
+
+import lazy_loader as lazy
+
+__getattr__, __dir__, __all__ = lazy.attach_stub(__name__, __file__)
